+++ wren-ai-service/src/pipelines/sql_generation/followup_generation.py
@@ -12,9 +12,9 @@ from langfuse.decorators import observe
 from src.core.engine import Engine
 from src.core.pipeline import BasicPipeline
 from src.core.provider import LLMProvider
-from src.pipelines.common import SQLGenerationPostProcessor
-from src.pipelines.sql_generation.components.prompts import (
+from src.pipelines.common import (
     TEXT_TO_SQL_RULES,
+    SQLGenerationPostProcessor,
     text_to_sql_system_prompt,
 )
 from src.utils import async_timer, timer
